Use icon on Android app
I was expecting to have to cross compile the Rust code before the Android app could actually go past the main menu, but just as well I checked because java.lang.foreign isn't actually supported on Android at all so nothing built from JExtract is going to work in the near term. Going to have to switch to a more mature solution but on the plus side this should mean it's feasible to lower the minSDK somewhat.