Remove error paths and handling now that ConfigHandle methods are non fallible
Could have probably done this if proceeding with JExtract too with unwind changes in Rust making it possible to abort the program instead of having to catch all panics at the boundry, but much less that can go wrong in general with more of the bindings glue being codegened. Especially way less work to pass strings over the boundry than with JExtract so less that can be written wrong as well.